Optimize NMS (#14290)
* Optimize NMS

* Fix lint
